Destroy is Finished

+++ app/Repositories/Admin/TransportationRepository.php
@@ -258,13 +258,17 @@ class TransportationRepository extends BaseRepository
     public function destroyOrderContract($request)
     {
         $transportOrder = TransportOrder::findOrFail($request->orderId);
-        $filePath = public_path($transportOrder->contract);
-        if (file_exists($filePath)) {
-            unlink($filePath);
+        $transportsUsingSameFile = TransportOrder::where('contract', $transportOrder->contract)
+            ->where('id', '!=', $request->orderId)
+            ->get();
+        if ($transportsUsingSameFile->isEmpty()) {
+            $filePath = public_path($transportOrder->contract);
+            if (file_exists($filePath)) {
+                unlink($filePath);
+            }
         }
         $transportOrder->last_price = null;
         $transportOrder->contract = null;
         $transportOrder->save();
-        return $request->orderId;
     }
 }
